feat: refactor page layout with TabBar navigation
- Add TabBar with three tabs: Home, Records, Settings
- Home page: app title and quick action buttons to add records
- Records page: date-based record viewer (moved from previous home page)
- Settings page: user profile with avatar and nickname

Closes #37
