+++ src/pages/index/index.css
@@ -1,7 +1,6 @@
-.overview-page {
+.home-page {
   min-height: 100vh;
   background-color: #f5f5f5;
-  padding-bottom: 40px;
 }
 
 /* 顶部栏 */
@@ -9,20 +8,31 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 24px 32px;
+  padding: 40px 32px;
   background-color: #fff;
 }
 
+.title-section {
+  display: flex;
+  flex-direction: column;
+}
+
 .app-title {
-  font-size: 40px;
+  font-size: 56px;
   font-weight: 700;
   color: #07c160;
+  margin-bottom: 8px;
+}
+
+.app-subtitle {
+  font-size: 28px;
+  color: #666;
 }
 
 .header-avatar {
-  width: 72px;
-  height: 72px;
-  border-radius: 36px;
+  width: 96px;
+  height: 96px;
+  border-radius: 48px;
   overflow: hidden;
   flex-shrink: 0;
 }
@@ -38,127 +48,43 @@
   background-color: #e0e0e0;
 }
 
-/* 日期选择器 */
-.date-selector {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 20px 32px;
-  background-color: #fff;
-  border-top: 1px solid #f0f0f0;
+/* 快速操作 */
+.quick-actions {
+  padding: 40px 32px;
+}
+
+.section-title {
+  font-size: 32px;
+  font-weight: 600;
+  color: #333;
+  margin-bottom: 32px;
+  display: block;
+}
+
+.action-grid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
   gap: 24px;
 }
 
-.date-arrow {
-  font-size: 36px;
-  color: #07c160;
-  padding: 16px;
-}
-
-.date-arrow.disabled {
-  color: #ccc;
-}
-
-.date-text {
-  font-size: 36px;
-  font-weight: 600;
-  color: #333;
-}
-
-/* 加载状态 */
-.loading {
+.action-item {
   display: flex;
-  justify-content: center;
+  flex-direction: column;
   align-items: center;
-  padding: 80px;
-  color: #999;
-  font-size: 28px;
-}
-
-/* 记录容器 */
-.records-container {
-  padding: 0 24px;
-}
-
-/* 记录卡片 */
-.record-card {
+  justify-content: center;
+  padding: 48px 24px;
   background-color: #fff;
   border-radius: 16px;
-  margin-bottom: 20px;
-  overflow: hidden;
+  box-shadow: 0 2px 8px rgb(0 0 0 / 5%);
 }
 
-.card-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 24px 28px;
-  border-bottom: 1px solid #f0f0f0;
+.action-icon {
+  font-size: 56px;
+  margin-bottom: 16px;
 }
 
-.card-title-row {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.card-icon {
-  font-size: 36px;
-}
-
-.card-title {
+.action-label {
   font-size: 30px;
-  font-weight: 600;
   color: #333;
-}
-
-.card-count {
-  font-size: 24px;
-  color: #999;
-}
-
-.card-add-btn {
-  font-size: 40px;
-  color: #07c160;
-  padding: 8px 16px;
-}
-
-.card-content {
-  padding: 20px 28px;
-  min-height: 60px;
-}
-
-.empty-hint {
-  font-size: 26px;
-  color: #ccc;
-  text-align: center;
-  display: block;
-  padding: 20px 0;
-}
-
-.record-item {
-  display: flex;
-  align-items: center;
-  padding: 12px 0;
-  gap: 16px;
-}
-
-.record-time {
-  font-size: 26px;
-  color: #999;
-  min-width: 90px;
-}
-
-.record-feeling {
-  font-size: 28px;
-  margin-right: 8px;
-}
-
-.record-desc {
-  font-size: 28px;
-  color: #333;
-  flex: 1;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
+  font-weight: 500;
 }
