+++ src/pages/settings/index.css
@@ -1,0 +1,55 @@
+.settings-page {
+  min-height: 100vh;
+  background-color: #f5f5f5;
+}
+
+/* 个人资料 */
+.profile-section {
+  display: flex;
+  align-items: center;
+  padding: 32px;
+  background-color: #fff;
+  margin-bottom: 20px;
+}
+
+.profile-avatar {
+  width: 120px;
+  height: 120px;
+  border-radius: 60px;
+  overflow: hidden;
+  flex-shrink: 0;
+  margin-right: 32px;
+}
+
+.avatar-img {
+  width: 100%;
+  height: 100%;
+}
+
+.avatar-default {
+  width: 100%;
+  height: 100%;
+  background-color: #e0e0e0;
+}
+
+.profile-info {
+  flex: 1;
+}
+
+.profile-nickname {
+  font-size: 36px;
+  font-weight: 600;
+  color: #333;
+  display: block;
+  margin-bottom: 8px;
+}
+
+.profile-hint {
+  font-size: 26px;
+  color: #999;
+}
+
+.profile-arrow {
+  font-size: 40px;
+  color: #ccc;
+}
